Fixed and extended skf file functionality
This commit fixes a bug in the Slater-Koster file parer that caused it to read one more row of the integral table than permitted. Bugs in the interpolation code related to this issue, which originally compensated for the previous off-by-one error, have also been resolved. Unit tests have been updated to account for the new changes.

A new method has been introduced to permit downloading of common Slater-Koster parameter sets from within TBMaLT.