Refactored DFTB repulsive feed (#86)
Refactored DFTB repulsive feed

This commit refactors the DFTB repulsive feed class "`RepulsiveSplineFeed`" to add support for periodic systems and to decouple the pair-wise interpolation functionality from the geometry level wrapper logic.

It should be noted that this breaks the `force` method of the `Dftb1` and `Dftb2` classes. A subsequent commit will be needed to resolve the issues present in the force method and remove the original "`RepulsiveSplineFeed`" class.