Simplified cutoff related functionality
The `cutoff` and `tail_distance` values have been unified so that only a single distance value, `cutoff`, needs to be provided to a `Geometry` instance.

The `positive_extension` and `negative_extension` values have been removed as they are no longer necessary due to changes in the `Periodicity.get_cell_translation_vector_indices` method.

The `Periodicity` class no longer permits units other than Bohr, and the `_check` method may no longer modify attributes of the `Geometry` instance that is creating it.

Documentation of the `Periodicity` class has been updated to indicate that instances should not be created manually. Rather this should be left to the `Geometry` class. At least until the recursion issue can be resolved.