fix: resolved several shape related bugs
Replaced several `torch.view` calls with their `torch.reshape` counterparts. This allows PyTorch to make a copy of a tensor instead of a view when it encounters a non-contiguous tensor rather than raising an exception.

The replaced `torch.unsqueeze` with a `torch.reshape` operation in the `gamma_exponential` method to avoid unintentionally striping the batch dimension from batches of size one.

Fixed the `merge_systems` helper function within several unit tests so that it correctly returns a batch of size one when invoked on a single system. Added a new method `Geometry.unsqueeze()` to allow for single system `Geometry` instances to be easily converted into batches of size one.